requirements.txt
streamlit
streamlit-bokeh-events==0.1.2
replicate
bokeh==2.4.3
pydub
simpleaudio
requests
python-dotenv
gTTS==2.3.1